# logs/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nodejs Training API",
    "version": "1.0.0",
    "description": "API for creating and managing users"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "Login to web",
        "description": null,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "itachi"
                  },
                  "password": {
                    "type": "string",
                    "example": "thisispassword"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successfully"
          },
          "400": {
            "description": "error message"
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Register a user",
        "description": null,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "UchihaItachi"
                  },
                  "name": {
                    "type": "string",
                    "example": "Uchiha Itachi"
                  },
                  "email": {
                    "type": "string",
                    "example": "itachi@uchiha.com"
                  },
                  "phone": {
                    "type": "string",
                    "example": "03367876621"
                  },
                  "password": {
                    "type": "string",
                    "example": "thisispassword@1"
                  }
                },
                "required": [
                  "name",
                  "email",
                  "phone",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "error message"
          }
        }
      }
    },
    "/event": {
      "get": {
        "summary": "Get events",
        "description": "This endpoint allows you to get events.",
        "tags": [
          "Event"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Event return"
          },
          "400": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create new event",
        "description": "This endpoint allows you to create new event.",
        "tags": [
          "Event"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "voucher_name": {
                    "type": "string",
                    "example": "Super sale 10.10"
                  },
                  "description": {
                    "type": "text",
                    "example": "what ever you want"
                  },
                  "voucher_date_start": {
                    "type": "string",
                    "example": "2024-10-01T01:25:18.375+00:00"
                  },
                  "voucher_date_end": {
                    "type": "string",
                    "example": "2024-10-11T01:25:18.375+00:00"
                  },
                  "voucher_quantity": {
                    "type": "number",
                    "example": "100"
                  },
                  "voucher_released": {
                    "type": "number",
                    "example": "100"
                  }
                },
                "required": [
                  "voucher_name",
                  "description",
                  "voucher_date_start",
                  "voucher_date_end",
                  "voucher_quantity",
                  "voucher_released"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event create successfully"
          },
          "456": {
            "description": "Some thing wrong"
          }
        }
      }
    },
    "/event/{id}": {
      "get": {
        "summary": "Edit a event",
        "description": "This endpoint allows you to edit a event.",
        "tags": [
          "Event"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Event return"
          },
          "400": {
            "description": "Event not found"
          }
        }
      },
      "put": {
        "summary": "Edit event",
        "description": "This endpoint allows you to edit event.",
        "tags": [
          "Event"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "voucher_name": {
                    "type": "string",
                    "example": "Super sale 10.10"
                  },
                  "description": {
                    "type": "text",
                    "example": "what ever you want"
                  },
                  "voucher_date_start": {
                    "type": "string",
                    "example": "2024-10-01T01:25:18.375+00:00"
                  },
                  "voucher_date_end": {
                    "type": "string",
                    "example": "2024-10-11T01:25:18.375+00:00"
                  },
                  "voucher_quantity": {
                    "type": "number",
                    "example": "100"
                  },
                  "voucher_released": {
                    "type": "number",
                    "example": "100"
                  }
                },
                "required": [
                  "voucher_name",
                  "description",
                  "voucher_date_start",
                  "voucher_date_end",
                  "voucher_quantity",
                  "voucher_released"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event update successfully"
          },
          "456": {
            "description": "Some thing wrong"
          }
        }
      },
      "delete": {
        "summary": "Delete event",
        "description": "This endpoint allows you to delete event.",
        "tags": [
          "Event"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Event delete successfully"
          },
          "456": {
            "description": "Some thing wrong"
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "This endpoint allows you to get all users",
        "description": "Get all users in user table",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Get users successfully"
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "description": "This endpoint allows you to create a new user.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Uchiha Itachi"
                  },
                  "email": {
                    "type": "string",
                    "example": "itachi@uchiha.com"
                  },
                  "phone": {
                    "type": "string",
                    "example": "03367876621"
                  },
                  "password": {
                    "type": "string",
                    "example": "thisispassword"
                  }
                },
                "required": [
                  "name",
                  "email",
                  "phone",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "error message"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "summary": "This endpoint allows you to get a user by id",
        "description": "Get user by id",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Get user successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "summary": "Edit a new user",
        "description": "This endpoint allows you to edit a new user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Uchiha Itachi"
                  },
                  "email": {
                    "type": "string",
                    "example": "itachi@uchiha.com"
                  },
                  "phone": {
                    "type": "string",
                    "example": "03367876621"
                  },
                  "password": {
                    "type": "string",
                    "example": "thisispassword"
                  }
                },
                "required": [
                  "name",
                  "email",
                  "phone",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User Edited successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "description": "This endpoint allows you to delete a user.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "User deleted successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/voucher": {
      "get": {
        "summary": "Get vouchers",
        "description": "This endpoint allows you to get vouchers.",
        "tags": [
          "Voucher"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Voucher return"
          },
          "400": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create new voucher",
        "description": "This endpoint allows you to create new voucher.",
        "tags": [
          "Voucher"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event_id": {
                    "type": "string",
                    "example": "1a212f32se41"
                  },
                  "event": {
                    "type": "Object",
                    "example": "{}"
                  },
                  "voucher_code": {
                    "type": "string",
                    "example": "Kute-1518"
                  },
                  "issued_to": {
                    "type": "string",
                    "example": "Johnson@gmailz.com"
                  },
                  "issued_date": {
                    "type": "Date",
                    "example": "2024-09-26T09:25:18.375+00:00"
                  },
                  "expired_date": {
                    "type": "Date",
                    "example": "2024-09-30T09:25:18.375+00:00"
                  }
                },
                "required": [
                  "event_id",
                  "event",
                  "voucher_code",
                  "issued_to",
                  "issued_date",
                  "expired_date"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Voucher create successfully"
          },
          "456": {
            "description": "Some thing wrong"
          }
        }
      }
    },
    "/voucher/{id}": {
      "get": {
        "summary": "Edit a voucher",
        "description": "This endpoint allows you to edit a voucher.",
        "tags": [
          "Voucher"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Voucher return"
          },
          "400": {
            "description": "Voucher not found"
          }
        }
      },
      "put": {
        "summary": "Edit voucher",
        "description": "This endpoint allows you to edit voucher.",
        "tags": [
          "Voucher"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event_id": {
                    "type": "string",
                    "example": "1a212f32se41"
                  },
                  "event": {
                    "type": "Object",
                    "example": "{}"
                  },
                  "voucher_code": {
                    "type": "string",
                    "example": "Kute-1518"
                  },
                  "issued_to": {
                    "type": "string",
                    "example": "Johnson@gmailz.com"
                  },
                  "issued_date": {
                    "type": "Date",
                    "example": "2024-09-26T09:25:18.375+00:00"
                  },
                  "expired_date": {
                    "type": "Date",
                    "example": "2024-09-30T09:25:18.375+00:00"
                  }
                },
                "required": [
                  "event_id",
                  "event",
                  "voucher_code",
                  "issued_to",
                  "issued_date",
                  "expired_date"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Voucher update successfully"
          },
          "456": {
            "description": "Some thing wrong"
          }
        }
      },
      "delete": {
        "summary": "Delete voucher",
        "description": "This endpoint allows you to delete voucher.",
        "tags": [
          "Voucher"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Voucher delete successfully"
          },
          "456": {
            "description": "Some thing wrong"
          }
        }
      }
    }
  },
  "tags": []
}